Don't assume Google Storage destination for videos
Inspect the job descrption instead.

( Which really needs to be done correctly--there should probably be an
'output_type' field in the job that could be examined (just like there's
an input type, generically named 'type' currently) as opposed to abusing
the 'output' field by stuffing different arbitrary strings into it and
then regexing the value to determine what it is.  This was just done to
quickly support outputting to local filesystem instead of assuming a
gstore bucket would be used.